Altro.jl API

Altro.jl API

Solvers

This page provides the docstrings for the most common methods that the user may work with.

Solvers

Types

Altro.ALTROSolver — Type

struct ALTROSolver{T, S} <: Altro.ConstrainedSolver{T}

Augmented Lagrangian Trajectory Optimizer (ALTRO) is a solver developed by the Robotic Exploration Lab at Stanford University. The solver is special-cased to solve Markov Decision Processes by leveraging the internal problem structure.

ALTRO consists of two "phases":

AL-iLQR: iLQR is used with an Augmented Lagrangian framework to solve the problem quickly to rough constraint satisfaction

Projected Newton: A collocation-flavored active-set solver projects the solution from AL-iLQR onto the feasible subspace to achieve machine-precision constraint satisfaction.

Constructor

ALTROSolver(prob::Problem, opts::SolverOptions; [infeasible, R_inf, kwarg_opts...])

The infeasible keyword is a boolean flag that specifies whether the solver should augment the controls to make it artificially fully actuated, allowing state initialization. The R_inf is the weight on these augmented controls. Any solver options can be passed as additional keyword arguments and will be set in the solver.

Getters

get_model

get_objective

get_constraints

Altro.get_ilqr

TO.get_initial_state

Other methods

Base.size: returns (n,m,N)

TO.integration

TO.is_constrained

Altro.AugmentedLagrangianSolver — Type

struct AugmentedLagrangianSolver <: ConstrainedSolver{T}

Augmented Lagrangian (AL) is a standard tool for constrained optimization. For a trajectory optimization problem of the form:

\[\begin{aligned} \min_{x_{0:N},u_{0:N-1}} \quad & \ell_f(x_N) + \sum_{k=0}^{N-1} \ell_k(x_k, u_k, dt) \\ \textrm{s.t.} \quad & x_{k+1} = f(x_k, u_k), \\ & g_k(x_k,u_k) \leq 0, \\ & h_k(x_k,u_k) = 0. \end{aligned}\]

AL methods form the following augmented Lagrangian function:

\[\begin{aligned} \ell_f(x_N) + &λ_N^T c_N(x_N) + c_N(x_N)^T I_{\mu_N} c_N(x_N) \\ & + \sum_{k=0}^{N-1} \ell_k(x_k,u_k,dt) + λ_k^T c_k(x_k,u_k) + c_k(x_k,u_k)^T I_{\mu_k} c_k(x_k,u_k) \end{aligned}\]

This function is then minimized with respect to the primal variables using any unconstrained minimization solver (e.g. iLQR). After a local minima is found, the AL method updates the Lagrange multipliers λ and the penalty terms μ and repeats the unconstrained minimization. AL methods have superlinear convergence as long as the penalty term μ is updated each iteration.

Altro.iLQRSolver — Type

iLQRSolver

A fast solver for unconstrained trajectory optimization that uses a Riccati recursion to solve for a local feedback controller around the current trajectory, and then simulates the system forward using the derived feedback control law.

Constructor

Altro.iLQRSolver(prob, opts; kwarg_opts...)

Altro.ProjectedNewtonSolver — Type

struct ProjectedNewtonSolver{T, N, M, NM} <: Altro.ConstrainedSolver{T}

Projected Newton Solver Direct method developed by the REx Lab at Stanford University Achieves machine-level constraint satisfaction by projecting onto the feasible subspace. It can also take a full Newton step by solving the KKT system. This solver is to be used exlusively for solutions that are close to the optimal solution. It is intended to be used as a "solution polishing" method for augmented Lagrangian methods.

Methods

Altro.set_options! — Function

set_options!(opts::AbstractSolverOptions; kwargs...)
set_options!(solver::AbstractSolver; opts...)

Set solver options via keyword arguments, supporting either solvers or solver option types directly. Will set any and all options that match the provided arguments. For example, set_options!(solver::ALTROSolver, constraint_tolerance=1e-4) will set the constraint tolerance option in the ALTRO, Augmented Lagrangian, and Project Newton solvers.

The only exeption is the verbose setting, which always accepts a boolean, while ALTRO and Augmented Lagrangian solvers accept integers 0-2, with 1 providing output for the outer AL iterations but not the iLQR iterations.

Altro.backwardpass! — Function

Calculates the optimal feedback gains K,d as well as the 2nd Order approximation of the Cost-to-Go, using a backward Riccati-style recursion. (non-allocating)

Altro.forwardpass! — Function

forwardpass!(solver, ΔV, J_prev)

Simulate the system forward using the optimal feedback gains from the backward pass, projecting the system on the dynamically feasible subspace. Performs a line search to ensure adequate progress on the nonlinear problem.

Altro.record_iteration! — Function

Stash iteration statistics

```